新しいコンポーネントとしてページネーションを作成
